Add config knobs and some code cleanup
