+++ cmd/server/server.go
@@ -137,6 +137,7 @@ func (s *server) startService() common.Daemon {
 		clusterMetadata.MasterClusterName,
 		clusterMetadata.CurrentClusterName,
 		clusterMetadata.ClusterInformation,
+		clusterMetadata.ReplicationConsumer,
 	)
 
 	if s.cfg.PublicClient.HostPort != "" {
