+++ common/cluster/metadata.go
@@ -49,6 +49,8 @@ type (
 		GetAllClusterInfo() map[string]config.ClusterInformation
 		// ClusterNameForFailoverVersion return the corresponding cluster name for a given failover version
 		ClusterNameForFailoverVersion(failoverVersion int64) string
+		// GetReplicationConsumerConfig returns the config for replication task consumer.
+		GetReplicationConsumerConfig() *config.ReplicationConsumerConfig
 	}
 
 	metadataImpl struct {
@@ -67,6 +69,8 @@ type (
 		clusterInfo map[string]config.ClusterInformation
 		// versionToClusterName contains all initial version -> corresponding cluster name
 		versionToClusterName map[int64]string
+		//replicationConsumer returns the config for replication task consumer.
+		replicationConsumer *config.ReplicationConsumerConfig
 	}
 )
 
@@ -78,6 +82,7 @@ func NewMetadata(
 	masterClusterName string,
 	currentClusterName string,
 	clusterInfo map[string]config.ClusterInformation,
+	replicationConsumer *config.ReplicationConsumerConfig,
 ) Metadata {
 
 	if len(clusterInfo) == 0 {
@@ -122,6 +127,7 @@ func NewMetadata(
 	return &metadataImpl{
 		logger:                   logger,
 		enableGlobalDomain:       enableGlobalDomain,
+		replicationConsumer:      replicationConsumer,
 		failoverVersionIncrement: failoverVersionIncrement,
 		masterClusterName:        masterClusterName,
 		currentClusterName:       currentClusterName,
@@ -194,4 +200,12 @@ func (metadata *metadataImpl) ClusterNameForFailoverVersion(failoverVersion int6
 		))
 	}
 	return clusterName
+}
+
+func (metadata *metadataImpl) GetReplicationConsumerConfig() *config.ReplicationConsumerConfig {
+	if metadata.replicationConsumer == nil {
+		return &config.ReplicationConsumerConfig{Type: config.ReplicationConsumerTypeKafka}
+	}
+
+	return metadata.replicationConsumer
 }
