+++ host/testcluster.go
@@ -107,6 +107,7 @@ func NewCluster(options *TestClusterConfig, logger log.Logger) (*TestCluster, er
 			options.ClusterMetadata.MasterClusterName,
 			options.ClusterMetadata.CurrentClusterName,
 			options.ClusterMetadata.ClusterInformation,
+			options.ClusterMetadata.ReplicationConsumer,
 		)
 	}
